Update README.md
#Project BlackJack
It is a Python code for the game BlackJack. For those, who are not aware of the game. 
It is a 2 player card game where number "21" represents BlackJack. 
In the game, each of the player is dealt two cards, with one card visible to us of the opponent and one hidden. If our score is below 17, we need to hit for more cards but at anytime, we exceed 21, we lose. If all the cards are dealt and the score of our opponent is less than us, we win and if not, we lose.